requirements.txt
absl-py==0.2.2
appnope==0.1.0
astor==0.7.1
atari-py==0.1.1
backcall==0.1.0
-e git+https://github.com/openai/baselines.git@f2729693253c0ef4d4086231d36e0a4307ec1cb3#egg=baselines
bleach==1.5.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
cycler==0.10.0
Cython==0.28.4
decorator==4.3.0
dill==0.2.8.2
entrypoints==0.2.3
future==0.16.0
gast==0.2.0
glfw==1.7.0
grpcio==1.13.0
-e git+https://github.com/StanfordASL/ALPaCA.git@27cd5e1a03fc255a2c559f1f187ea841fb71936b#egg=gym&subdirectory=gym
h5py==2.8.0
html5lib==0.9999999
idna==2.7
imageio==2.3.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.1
Jinja2==2.10
joblib==0.12.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mpi4py==3.0.0
mujoco-py==1.50.1.56
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.5
opencv-python==3.4.2.16
pandas==0.23.3
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
progressbar2==3.38.0
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
pycparser==2.18
pyglet==1.3.2
Pygments==2.2.0
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.5
PyYAML==3.13
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.20.0
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
tqdm==4.23.4
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.1
zmq==0.0.0
